Connected signup to database
