+++ App.py
@@ -26,6 +26,50 @@ def check_credentials():
 
     return jsonify({'authenticated': False})
 
+def encrypt(inputText, N, D):
+    # Check if N is >= 1
+    if N < 1:
+        return "Invalid value for N. N must be >= 1."
+
+    # Check if D is either 1 (shift right) or -1 (shift left)
+    if D not in [-1, 1]:
+        return "Invalid value for D. D must be either 1 (shift right) or -1 (shift left)."
+
+    encryptedText = ""
+    for char in reversed(inputText):
+        # If character is 'space' or '!', add it to the encrypted text without encryption
+        if char in [' ', '!']:
+            encryptedText += char
+            continue
+        
+        # Shift the character by N positions in the specified direction
+        shifted_char = chr(ord(char) + N * D)
+        encryptedText += shifted_char
+
+    return encryptedText
+
+def decrypt(encryptedText, N, D):
+    # Check if N is >= 1
+    if N < 1:
+        return "Invalid value for N. N must be >= 1."
+
+    # Check if D is either 1 (shift right) or -1 (shift left)
+    if D not in [-1, 1]:
+        return "Invalid value for D. D must be either 1 (shift right) or -1 (shift left)."
+
+    decryptedText = ""
+    for char in reversed(encryptedText):
+        # If character is 'space' or '!', add it to the decrypted text without decryption
+        if char in [' ', '!']:
+            decryptedText += char
+            continue
+        
+        # Shift the character by N positions in the opposite direction
+        shifted_char = chr(ord(char) - N * D)
+        decryptedText += shifted_char
+
+    return decryptedText
+
 @app.route('/signup', methods=['POST'])
 def signup():
     data = request.json
@@ -34,15 +78,15 @@ def signup():
     newUserID = data.get('newUserID')  
     newPassword = data.get('newPassword')
 
+    encrypted_ID = encrypt(newUserID, 3, 1)
+    encrypted_password = encrypt(newPassword, 3, 1)
+
     # Check if user already exists
-    if Database.user_exists(newUserID):
+    if Database.user_exists(encrypted_ID):
         return jsonify({"error": "User already exists", "code": 409}), 409
 
-    # Encrypt the password
-    encrypted_password = main.encrypt(newPassword)
-
     # Create the user in the database
-    Database.create_user(username, newUserID, encrypted_password)
+    Database.create_user(username, encrypted_ID, encrypted_password)
 
     return jsonify({"message": "User created successfully", "code": 200}), 200
